Major code update
So I just realised that I hadn't pushed code for quite a while.
There was a major refactoring of the data layer which also required
changes in the UI layer and I wasn't sure I wanted to go with it until
it was working.

Now that I'm happy with the code it's too hard to go through the changes
and document them properly, so I'm just going to dump it and hope
nobody notices/cares.